+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to benchkit will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.3] - 2024-06-15
+
+- Option in docker builders to use or not the docker buildkit;
+- Optional passing of UID/GID for the docker user;
+- Method to get a concrete docker runner from docker builder;
+- DockerfileDockerBuilder: new kind of docker builder that builds the image from a given dockerfile.
+
 ## [0.0.2] - 2024-04-25
 
 Updated example in README.